created app folder that holds glsl setup and gl draw codes. shader utils moved to same directory.
