+++ src/app/01_triangle/triangle.h
@@ -1,0 +1,7 @@
+#ifndef TRIANGLE_H
+#define TRIANGLE_H
+
+int init_triangle();
+void draw_triangle();
+
+#endif
