+++ src/common/common.h
@@ -122,7 +122,6 @@ static inline int __egl_check(void *ptr, const char *name)
 
 int init_egl(struct egl *egl, const struct gbm *gbm, int samples);
 int create_program(const char *vs_src, const char *fs_src);
-int link_program(unsigned program);
 
 enum mode {
 	SMOOTH,        /* smooth-shaded */
