+++ src/renderer.h
@@ -17,8 +17,4 @@ struct render_context {
 int setup_rendering(struct drm *drm, struct render_context *ctx);
 void cleanup_rendering(struct render_context *ctx);
 
-// New functions to initialize and draw a triangle
-int init_triangle();
-void draw_triangle();
-
 #endif
